+++ lib/src/helper.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 
 class CustomText extends StatefulWidget {
   const CustomText(
-      {Key key,
+      {Key? key,
       this.msg,
       this.style,
       this.textAlign,
@@ -12,18 +12,18 @@ class CustomText extends StatefulWidget {
       this.softwrap})
       : super(key: key);
 
-  final BuildContext context;
-  final String msg;
-  final TextOverflow overflow;
-  final bool softwrap;
-  final TextStyle style;
-  final TextAlign textAlign;
+  final BuildContext? context;
+  final String? msg;
+  final TextOverflow? overflow;
+  final bool? softwrap;
+  final TextStyle? style;
+  final TextAlign? textAlign;
 
   _CustomTextState createState() => _CustomTextState();
 }
 
 class _CustomTextState extends State<CustomText> {
-  TextStyle style;
+  TextStyle? style;
 
   @override
   @override
@@ -37,15 +37,15 @@ class _CustomTextState extends State<CustomText> {
       return Container();
     }
     if (widget.context != null && widget.style != null) {
-      var font = widget.style.fontSize == null
-          ? Theme.of(context).textTheme.bodyText1.fontSize
-          : widget.style.fontSize;
-      style = widget.style.copyWith(
+      var font = widget.style!.fontSize == null
+          ? Theme.of(context).textTheme.bodyText1!.fontSize!
+          : widget.style!.fontSize!;
+      style = widget.style!.copyWith(
           fontSize:
               font - (EmptyWidgetUtility.fullWidth(context) <= 375 ? 2 : 0));
     }
     return Text(
-      widget.msg,
+      widget.msg!,
       style: widget.style,
       textAlign: widget.textAlign,
       overflow: widget.overflow,